requirements.txt
langchain-openai
langchain_anthropic
langchain_cohere
langchain-community
langchain-core
langgraph
tavily-python
python-dotenv
python-docx


